change -language flag to -game flag
Though there's not supposed to be anything wrong with using `-language`, ColorzCore does not recognize it and as such treats every raw that uses it instead of `-game` as being valid regardless of the game setting. This most commonly breaks ASMCs when the first ASMC raw encountered is the FE6 one. Only a handful of raws used `-language`, so switching them to `-game` also establishes unified formatting.